Wrap the content of todotestdb.sql into a procedure.
Now, if the whole file is copypasted, a procedure is created for
creating the example database. This procedure is then called.
This makes resetting the example db much easier, as you only
need to call the procedure and then you'll get a fresh example
db.
Also add new dummy tasks.